Update MiscViewModel.cs
changed _tSearchGameLinks and _tSearchGameNames to observable collection. Not sure if this causes any issues, but this change was to make the UI refresh when selection of DB changed

implemented IGDB logic in view model. 

added new function to clear list in UI when option is changed. I had to add this for the stupid users who will complain when they search for games on IGDB, then switch to TA, and then press on the game they searched and get an error.